Renamed `BindSensorViewModel` to `BindDialogViewModelImpl`
