+++ feature/unlocated/src/main/kotlin/com/masselis/tpmsadvanced/unlocated/ioc/FeatureUnlocatedBinding.kt
@@ -3,11 +3,11 @@ package com.masselis.tpmsadvanced.unlocated.ioc
 import com.masselis.tpmsadvanced.core.feature.ioc.FeatureCoreComponent
 import com.masselis.tpmsadvanced.data.unit.ioc.DataUnitComponent
 import com.masselis.tpmsadvanced.data.vehicle.ioc.DataVehicleComponent
-import com.masselis.tpmsadvanced.unlocated.interfaces.viewmodel.BindSensorViewModelImpl
+import com.masselis.tpmsadvanced.unlocated.interfaces.viewmodel.BindDialogViewModelImpl
 import com.masselis.tpmsadvanced.unlocated.interfaces.viewmodel.ListSensorViewModelImpl
 import dagger.Component
 
-@Suppress("PropertyName", "FunctionName")
+@Suppress("PropertyName", "VariableNaming")
 @FeatureUnlocatedBinding.Scope
 @Component(
     dependencies = [
@@ -21,12 +21,8 @@ internal interface FeatureUnlocatedBinding {
     @javax.inject.Scope
     annotation class Scope
 
-
-    @Suppress("VariableNaming")
     val ListSensorViewModel: ListSensorViewModelImpl.Factory
-
-    @Suppress("VariableNaming")
-    val BindSensorViewModel: BindSensorViewModelImpl.Factory
+    val BindDialogViewModel: BindDialogViewModelImpl.Factory
 
     companion object : FeatureUnlocatedBinding by DaggerFeatureUnlocatedBinding
         .builder()
